gdkdisplay-win32.c: Reduce usage on _gdk_display
For the hidden window that we create to register display change
notifications from Windows, pass in the GdkDisplay that we have into the
CreateWindow() call for creating the hidden window (HWND) to receive the
system messages when for instance, display resolution is changed.

This way, we can save it up in the user data area in the HWND and
retrieve it later when we handle WM_DISPLAYCHANGE upon a display
resolution change, instead of using a global _gdk_display in this case.