gdkdevicemanager-win32.c: Stop using _gdk_display
Instead, add a GObject property for the GdkDisplay in
GdkDeviceManagerWin32 and pass it into the g_object_new() call when we
create the GdkDeviceManagerWin32 object during initialization.

Also, for the keyboard and pointer (mouse, etc) device creation, change
the function calls a bit as we aren't using the GdkDeviceManagerWin32
object for them, but instead we are using the GdkDisplay.

For now, we will still use _gdk_display when creating the object, but
only from gdkdisplay-win32.c.