gdkinput-winpointer.c: Drop _gdk_display usage
For the hidden window that we create to receive WM_POINTERDEVICECHANGE
notifications from Windows, pass in the GdkDisplay that we have into the
CreateWindowExW() call for creating the hidden window (HWND) to receive
the system messages.

This way, we can save it up in the user data area in the HWND and
retrieve it later when we handle WM_POINTERDEVICECHANGE upon a touch
monitor settings change, instead of using a global _gdk_display in this
case.