gdkglcontext.c: Re-create a OpenGL/ES 3.0 context if needed
Currently, we create an OpenGL/ES 2.0 context if GLES is requested or required,
but it may not support the needed extensions for it to work, such as
OES_vertex_half_float.  So, if that is the case, create an OpenGL/ES 3.0
context to replace the OpenGL/ES 2.0 context that we have created, so that we
have the needed OpenGL/ES features, in which the OpenGL/ES 2.0 context will
then be dropped.