gridview: Make focus movement section-aware
Make up-down focus movements preserve the column
across section boundaries.

This code does the simplest thing that can work.
If this turns out to be a problem, it can be
made smarter.

Note that the code currently prefers to jump over
short sections to preserve the column. If there
are many short sections, this not be desired,
and we may want to make it tweakable.